fixed rust size exceeded
